getdeps: delete getdeps build for sapling (#4404)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4404

We are now planning to remove the Sapling getdeps build since Sapling
doesn't use it for open source anyway. The EdenFS getdeps build will adapt one
way or another.

The getdeps was added in D22336485, it seems there are more things to clean up,
this diff  is for deleting the getdeps build job

Differential Revision: D44350947

fbshipit-source-id: 6b0ba196b83529a2827702849511c65e7552486a